Rename `test` package into `testing`
