+++ buildSrc/src/main/kotlin/io/spine/internal/gradle/testing/Tasks.kt
@@ -24,7 +24,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.internal.gradle.test
+package io.spine.internal.gradle.testing
 
 import org.gradle.api.tasks.TaskContainer
 import org.gradle.api.tasks.testing.Test
